+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+* fix: Add `ignoreMissing` flag for `setCommit` command to JS API (#974)
+
 ## sentry-cli 1.65.0
 
 * feat: Allow for ignoring missing commits in set-commit with `--ignore-missing` flag (#963)